+++ HISTORY.rst
@@ -3,9 +3,11 @@ History
 
 Next Release
 ------------
+
 * Add a programmatic API in module ``memote.suite.api`` (#162).
 * Reorganize the structure and build process for auto-documenting ``memote`` (#172).
 * Add a new command ``memote online`` (#95, #153).
+* Add more basic tests.
 
 0.3.6 (2017-08-15)
 ------------------